+++ .github/scripts/compile_zip.sh
@@ -43,7 +43,7 @@ cp LICENSE ./module
 find module/system/bin -maxdepth 1 -type f -name "*.sh" -exec sh -c 'mv -- "$0" "${0%.sh}"' {} \;
 
 # Parse version info to module prop
-zipName="encore-$version-$release_code.zip"
+zipName="EronX_Tweaks-$version-$release_code.zip"
 echo "zipName=$zipName" >>"$GITHUB_OUTPUT"
 
 # Generate sha256sum for integrity checkup