+++ jni/src/EronX_Tweaks_profiler.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024-2025 Rem01Gaming
+ * Copyright (C) 2024-2025 ERON
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <encore.h>
+#include <EronX_Tweaks.h>
 
 bool (*get_screenstate)(void) = get_screenstate_normal;
 bool (*get_low_power_state)(void) = get_low_power_state_normal;
@@ -38,8 +38,8 @@ void run_profiler(const int profile) {
     }
 
     write2file(PROFILE_MODE, false, false, "%d\n", profile);
-    if (systemv("encore_profiler %d", profile)) {
-        log_encore(LOG_ERROR, "Unable to execute profiler changes to %d", profile);
+    if (systemv("EronX_Tweaks_profiler %d", profile)) {
+        log_EronX_Tweaks(LOG_ERROR, "Unable to execute profiler changes to %d", profile);
     }
 }
 
@@ -80,10 +80,10 @@ bool get_screenstate_normal(void) {
     }
 
     fetch_failed++;
-    log_encore(LOG_ERROR, "Unable to fetch current screenstate");
+    log_EronX_Tweaks(LOG_ERROR, "Unable to fetch current screenstate");
 
     if (fetch_failed == 6) {
-        log_encore(LOG_FATAL, "get_screenstate is out of order!");
+        log_EronX_Tweaks(LOG_FATAL, "get_screenstate is out of order!");
 
         // Set default state after too many failures via function pointer
         get_screenstate = return_true;
@@ -119,10 +119,10 @@ bool get_low_power_state_normal(void) {
     }
 
     fetch_failed++;
-    log_encore(LOG_ERROR, "Unable to fetch battery saver status");
+    log_EronX_Tweaks(LOG_ERROR, "Unable to fetch battery saver status");
 
     if (fetch_failed == 6) {
-        log_encore(LOG_FATAL, "get_low_power_state is out of order!");
+        log_EronX_Tweaks(LOG_FATAL, "get_low_power_state is out of order!");
 
         // Set default state after too many failures via function pointer
         get_low_power_state = return_false;