+++ changelog.md
@@ -1,6 +1,6 @@
-## Encore Tweaks 4.6
+## EronX Tweaks 4.6
 
-**Thanks for using Encore Tweaks!**
+**Thanks for using EronX Tweaks!**
 
 Your continued support keeps this project going. If you enjoy the improvements and want to see more features in the future, [consider supporting the developer with a small donation](https://t.me/rem01schannel/670). Every bit helps!
 
@@ -9,6 +9,6 @@ Your continued support keeps this project going. If you enjoy the improvements a
 - Refactor and design changes on the WebUI
 - Implement WebUI X API for file read/write
 - WebUI translation updates
-- Exposes Encore's logging system to other processes/module
+- Exposes EronX's logging system to other processes/module
 - Add sanity check for dumpsys
 - Other misc changes and optimizations